Fix implementation

+++ src/main/java/capgemini/lecture1/Calculator.java
@@ -3,8 +3,8 @@ package capgemini.lecture1;
 //Thread-safe
 public class Calculator {
 
-    public int sum(int firstOp, int secondOp) {
-        return firstOp + secondOp;
+    public long sum(int firstOp, int secondOp) {
+        return (long) firstOp + secondOp;
     }
